+++ R/utils.R
@@ -102,7 +102,7 @@ consistency_check <- function(obj, case, ...) {
                valid <- c("dtw", "dtw2", "sbd")
 
                ## pr_DB$entry_exists is acting weird
-               if (!is.character(obj) || class(try(pr_DB[[obj]], TRUE)) == "try-error") {
+               if (!is.character(obj) || !pr_DB$entry_exists(obj)) {
                     if (silent)
                          return(FALSE)
                     else
@@ -213,6 +213,13 @@ colMedians <- function(mat) { apply(mat, 2L, stats::median) }
 # PREFUN for some of my proxy distances so that they support 'pairwise' direclty
 # Currently results in warnings because proxy does !is.na(reg_entry$PREFUN)
 proxy_prefun <- function(x, y, pairwise, params, reg_entry) {
+     if (!is.null(params$force.pairwise)) {
+          warning("The argument 'force.pairwise. has been deprecated. Simply use 'pairwise' instead.")
+
+          pairwise <- params$force.pairwise
+          params$force.pairwise <- NULL
+     }
+
      params$pairwise <- pairwise
 
      list(x = x, y = y,