+++ tests/testthat/test_pairwise.R
@@ -4,7 +4,7 @@ context("Test pairwise")
 # sbd
 # =================================================================================================
 
-pdist_sbd <- proxy::dist(data[1:50], data[51:100], method = "SBD", force.pairwise = TRUE)
+pdist_sbd <- proxy::dist(data[1:50], data[51:100], method = "SBD", pairwise = TRUE)
 
 test_that("pairwise sbd distance gives the same result as reference",
           my_expect_equal_to_reference(pdist_sbd))
@@ -14,7 +14,7 @@ test_that("pairwise sbd distance gives the same result as reference",
 # =================================================================================================
 
 pdist_lbk <- proxy::dist(data_list[1:50], data_list[51:100], method = "LBK",
-                         window.size = 18L, force.pairwise = TRUE)
+                         window.size = 18L, pairwise = TRUE)
 
 test_that("pairwise lbk distance gives the same result as reference",
           my_expect_equal_to_reference(pdist_lbk))
@@ -24,7 +24,7 @@ test_that("pairwise lbk distance gives the same result as reference",
 # =================================================================================================
 
 pdist_lbi <- proxy::dist(data_list[1:50], data_list[51:100], method = "LBI",
-                         window.size = 18L, force.pairwise = TRUE)
+                         window.size = 18L, pairwise = TRUE)
 
 test_that("pairwise lbi distance gives the same result as reference",
           my_expect_equal_to_reference(pdist_lbi))
@@ -34,7 +34,7 @@ test_that("pairwise lbi distance gives the same result as reference",
 # =================================================================================================
 
 pdist_dtw_lb <- proxy::dist(data_list[1:50], data_list[51:100], method = "DTW_LB",
-                            window.size = 18L, force.pairwise = TRUE)
+                            window.size = 18L, pairwise = TRUE)
 
 test_that("pairwise dtw_lb distance gives the same result as reference",
           my_expect_equal_to_reference(pdist_dtw_lb))