+++ NEWS.md
@@ -1,7 +1,7 @@
 # NEWS
 
 ## Version 2.1.3
-* TBA
+* The custom argument `force.pairwise` used in some of the `proxy` distances is not necessary anymore.
 
 ## Version 2.1.2
 * Added option to specify a function to extract prototypes in hierarchical and TADPole clustering